# spec/openapi.yaml
openapi: 3.0.1
info:
  title: ebal API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /services:
    get:
      tags: [Services]
      operationId: listServices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Services page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageServiceResponse'
    post:
      tags: [Services]
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
  /services/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Services]
      operationId: getService
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
    put:
      tags: [Services]
      operationId: updateService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
    delete:
      tags: [Services]
      operationId: deleteService
      responses:
        '204':
          description: No Content
  /services/{id}/plan-items:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Services]
      operationId: listServicePlanItems
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicePlanItemResponse'
    post:
      tags: [Services]
      operationId: addServicePlanItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicePlanItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePlanItemResponse'
...
components:
  schemas:
...
    ServiceRequest:
      type: object
      required: [startsAt]
      properties:
        startsAt:
          type: string
          format: date-time
        location:
          type: string
    ServiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        location:
          type: string
    PageServiceResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ServiceResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
    ServicePlanItemRequest:
      type: object
      properties:
        type:
          type: string
        refId:
          type: string
          format: uuid
        sortOrder:
          type: integer
        notes:
          type: string
    ServicePlanItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        type:
          type: string
        refId:
          type: string
          format: uuid
        sortOrder:
          type: integer
        notes:
          type: string
...